# server/docs/swagger.yaml
basePath: /Student_Management_System
definitions:
  activity_handler.addResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  activity_handler.deleteResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  activity_handler.getInfoResponse:
    properties:
      activities:
        items:
          $ref: '#/definitions/system.ActivityInfo'
        type: array
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  common.LoginRequest:
    properties:
      password:
        type: string
      username:
        type: string
    type: object
  common.RegisterRequest:
    properties:
      password:
        type: string
      student_id:
        type: string
      username:
        type: string
    type: object
  course_handler.addResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  course_handler.deleteResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  course_handler.searchResponse:
    properties:
      course_list:
        items:
          $ref: '#/definitions/system.CourseInfo'
        type: array
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  course_handler.tableResponse:
    properties:
      course_list:
        items:
          $ref: '#/definitions/system.CourseInfo'
        type: array
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  course_handler.updateResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  log_handler.addResponse:
    properties:
      log_id:
        type: integer
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  log_handler.deleteResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  log_handler.infoResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/system.LogInfo'
        type: array
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  navigate_handler.NavigateResponse:
    properties:
      node_list:
        items:
          items:
            type: number
          type: array
        type: array
      path_list:
        items:
          $ref: '#/definitions/system.PathGuider'
        type: array
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  navigate_handler.tspResponse:
    properties:
      node_list:
        description: 途径节点的经纬度
        items:
          items:
            type: number
          type: array
        type: array
      pass_list:
        description: 用户选择的目标点的经纬度，用户在地图上显示
        items:
          items:
            type: number
          type: array
        type: array
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  polling_handler.isCourseArriveResponse:
    properties:
      course_info:
        $ref: '#/definitions/system.CourseInfo'
      is_arrive:
        type: boolean
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
  system.ActivityInfo:
    properties:
      activity_id:
        type: integer
      activity_name:
        type: string
      frequency:
        type: integer
      is_mention:
        description: 是否需要提醒
        type: boolean
      location:
        type: string
      start_time:
        type: string
      tag:
        type: string
      type:
        type: integer
      user_id:
        type: integer
    type: object
  system.CourseInfo:
    properties:
      classroom:
        description: 线下课程记录教室, 线上课程记录会议链接
        type: string
      contact:
        type: string
      course_id:
        type: integer
      course_location_node:
        description: 对应导航地图的节点, 设置成数组是因为一个地点可能对应多个node
        items:
          type: integer
        type: array
      course_name:
        type: string
      exam_location:
        type: string
      exam_location_node:
        items:
          type: integer
        type: array
      exam_option:
        type: integer
      exam_time:
        description: 考试时间, 格式为'第几周-星期几-几点~几点（HH:mm~HH:mm）', 如'第十六周-星期三-19:00~20:00'
        type: string
      is_compulsory:
        description: 课程是否必修
        type: boolean
      is_course_online:
        type: boolean
      section_list:
        items:
          type: integer
        type: array
      teacher:
        type: string
      week_schedule:
        items:
          type: integer
        type: array
    type: object
  system.LogInfo:
    properties:
      content:
        type: string
      create_time:
        type: string
      log_id:
        type: integer
      user_id:
        description: 对应的用户id, 对应某个用户的日志
        type: integer
    type: object
  system.PathGuider:
    properties:
      distance:
        description: 向量长度,单位为米
        type: number
      heading:
        description: 向量方向,用航向角来计算(正北为0,360等分)
        type: integer
    type: object
  system.UserInfo:
    properties:
      avatar_url:
        description: 保存用户头像的路径
        type: string
      is_admin:
        type: boolean
      signature:
        description: '......后续应该还要加上一些字段'
        type: string
      student_id:
        type: string
      user_id:
        type: integer
      username:
        type: string
    type: object
  user_handler.infoResponse:
    properties:
      status_code:
        type: integer
      status_msg:
        type: string
      user_info:
        $ref: '#/definitions/system.UserInfo'
    type: object
  user_handler.loginResponse:
    properties:
      is_admin:
        type: boolean
      status_code:
        type: integer
      status_msg:
        type: string
      token:
        type: string
      user_id:
        type: integer
    type: object
  user_handler.registerResponse:
    properties:
      is_admin:
        type: boolean
      status_code:
        type: integer
      status_msg:
        type: string
      token:
        type: string
      user_id:
        type: integer
    type: object
  user_handler.uploadAvatarResponse:
    properties:
      avatar_url:
        type: string
      status_code:
        type: integer
      status_msg:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A application for Data Structure Lesson
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Student Management System API
  version: alpha 1.0
paths:
  /activity/add:
    post:
      description: 增加活动信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/activity_handler.addResponse'
      security:
      - ApiKeyAuth: []
      summary: 增加活动信息
      tags:
      - 活动接口
  /activity/delete:
    delete:
      description: 删除活动信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/activity_handler.deleteResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除活动信息
      tags:
      - 活动接口
  /activity/info:
    get:
      description: 获取学生所有的课外活动信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/activity_handler.getInfoResponse'
      security:
      - ApiKeyAuth: []
      summary: 课外活动信息
      tags:
      - 活动接口
  /course/add:
    post:
      description: 管理员添加课程
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course_handler.addResponse'
      security:
      - ApiKeyAuth: []
      summary: 添加课程
      tags:
      - 课程接口
  /course/delete:
    delete:
      description: 根据id删除课程
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course_handler.deleteResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除课程
      tags:
      - 课程接口
  /course/search:
    get:
      description: 查询课程
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course_handler.searchResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询课程
      tags:
      - 课程接口
  /course/table:
    get:
      description: 获取学生所有的课程表信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course_handler.tableResponse'
      security:
      - ApiKeyAuth: []
      summary: 课程表
      tags:
      - 课程接口
  /course/update:
    put:
      description: 更新课程信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/course_handler.updateResponse'
      security:
      - ApiKeyAuth: []
      summary: 更新课程
      tags:
      - 课程接口
  /log/add:
    post:
      consumes:
      - application/json
      description: 添加一条日志
      parameters:
      - description: 时间
        in: query
        name: create_time
        required: true
        type: string
      - description: 内容
        in: query
        name: content
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/log_handler.addResponse'
      security:
      - ApiKeyAuth: []
      summary: 添加日志
      tags:
      - 日志接口
  /log/delete:
    get:
      consumes:
      - application/json
      description: 删除一条日志
      parameters:
      - description: 日志id
        in: query
        name: log_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/log_handler.deleteResponse'
      security:
      - ApiKeyAuth: []
      summary: 删除日志
      tags:
      - 日志接口
  /log/info:
    get:
      consumes:
      - application/json
      description: 查询日志
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/log_handler.infoResponse'
      security:
      - ApiKeyAuth: []
      summary: 查询日志
      tags:
      - 日志接口
  /navigate/path:
    get:
      consumes:
      - application/json
      description: 获取导航路线
      parameters:
      - description: 起始节点ID
        in: query
        name: from_id
        required: true
        type: integer
      - description: 结束节点ID
        in: query
        name: des_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/navigate_handler.NavigateResponse'
      security:
      - ApiKeyAuth: []
      summary: 导航
      tags:
      - 导航接口
  /navigate/tsp:
    get:
      consumes:
      - application/json
      description: 解决旅行商问题
      parameters:
      - description: 开始节点ID
        in: query
        name: start_id
        required: true
        type: integer
      - collectionFormat: csv
        description: 途径节点集合
        in: query
        items:
          type: integer
        name: des_ids
        required: true
        type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/navigate_handler.tspResponse'
      security:
      - ApiKeyAuth: []
      summary: 旅行商
      tags:
      - 导航接口
  /polling/is_course_arrive:
    get:
      consumes:
      - application/json
      description: 课程是否来临
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/polling_handler.isCourseArriveResponse'
      summary: 课程是否来临
      tags:
      - 轮询接口
  /user/info:
    get:
      consumes:
      - application/json
      description: 获取用户信息
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_handler.infoResponse'
      summary: 用户信息
      tags:
      - 用户接口
  /user/login:
    post:
      consumes:
      - application/json
      description: 用户登录
      parameters:
      - description: 用户信息
        in: body
        name: user_input
        required: true
        schema:
          $ref: '#/definitions/common.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_handler.loginResponse'
      summary: 登录
      tags:
      - 用户接口
  /user/register:
    post:
      consumes:
      - application/json
      description: 用户注册
      parameters:
      - description: 用户信息
        in: body
        name: user_input
        required: true
        schema:
          $ref: '#/definitions/common.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_handler.registerResponse'
      summary: 注册
      tags:
      - 用户接口
  /user/upload_avatar:
    post:
      consumes:
      - application/json
      description: 用户上传头像
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/user_handler.uploadAvatarResponse'
      summary: 上传头像
      tags:
      - 用户接口
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: token
    type: apiKey
swagger: "2.0"